+++ test/pull-request.test.ts
@@ -855,12 +855,22 @@ describe("Error scenarios", () => {
 
   test("blob size is too large", async () => {
     githubAPI.getPrFile.mockImplementationOnce(() => { throw new Error("Blob too large"); });
-    githubAPI.getPrNumber.mockReturnValueOnce(PR_NUMBER.ADD_SVG);
+    githubAPI.getPrNumber.mockReturnValueOnce(PR_NUMBER.MANY_CHANGES);
 
     await main(client, config, svgo);
 
     expect(core.setFailed).toHaveBeenCalledTimes(0);
-    expect(core.warning).toHaveBeenCalledWith(expect.stringContaining("SVG content could not be obtained"));
+    expect(core.warning).toHaveBeenCalledWith(expect.stringMatching("SVG content .*could not be obtained"));
+
+    expect(templating.formatCommitMessage).toHaveBeenCalledWith(
+      expect.any(String),
+      expect.any(String),
+      expect.objectContaining({
+        warnings: expect.arrayContaining([
+          expect.stringMatching("SVG content .*could not be obtained"),
+        ]),
+      }),
+    );
   });
 
   test("optimized blob size is too large", async () => {
@@ -870,7 +880,17 @@ describe("Error scenarios", () => {
     await main(client, config, svgo);
 
     expect(core.setFailed).toHaveBeenCalledTimes(0);
-    expect(core.warning).toHaveBeenCalledWith(expect.stringContaining("Blob could not be created"));
+    expect(core.warning).toHaveBeenCalledWith(expect.stringMatching("Blob.* could not be created"));
+
+    expect(templating.formatCommitMessage).toHaveBeenCalledWith(
+      expect.any(String),
+      expect.any(String),
+      expect.objectContaining({
+        warnings: expect.arrayContaining([
+          expect.stringMatching("Blob.* could not be created"),
+        ]),
+      }),
+    );
   });
 
   test("missing head reference", async () => {